Extracted BoboValueAnnotation tests to separate class
